+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/GoldService.scala
@@ -852,14 +852,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     type FunctionType = Function1[Args,Future[com.twitter.scrooge.test.gold.thriftscala.Response]]
     type ReqRepFunctionType = Function1[_root_.com.twitter.scrooge.Request[Args],Future[_root_.com.twitter.scrooge.Response[com.twitter.scrooge.test.gold.thriftscala.Response]]]
 
-    type ServiceIfaceServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ReqRepServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[_root_.com.twitter.scrooge.Request[Args], _root_.com.twitter.scrooge.Response[SuccessType]]
-
-    def toServiceIfaceService(f: FunctionType): ServiceIfaceServiceType =
-      _root_.com.twitter.finagle.Service.mk { args: Args =>
-        f(args)
-      }
 
     def toServicePerEndpointService(f: FunctionType): ServicePerEndpointServiceType =
       _root_.com.twitter.finagle.Service.mk { args: Args =>
@@ -1377,14 +1371,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     type FunctionType = Function1[Args,Future[com.twitter.scrooge.test.gold.thriftscala.Response]]
     type ReqRepFunctionType = Function1[_root_.com.twitter.scrooge.Request[Args],Future[_root_.com.twitter.scrooge.Response[com.twitter.scrooge.test.gold.thriftscala.Response]]]
 
-    type ServiceIfaceServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ReqRepServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[_root_.com.twitter.scrooge.Request[Args], _root_.com.twitter.scrooge.Response[SuccessType]]
-
-    def toServiceIfaceService(f: FunctionType): ServiceIfaceServiceType =
-      _root_.com.twitter.finagle.Service.mk { args: Args =>
-        f(args)
-      }
 
     def toServicePerEndpointService(f: FunctionType): ServicePerEndpointServiceType =
       _root_.com.twitter.finagle.Service.mk { args: Args =>