+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/PlatinumService.scala
@@ -952,14 +952,8 @@ object PlatinumService extends _root_.com.twitter.finagle.thrift.GeneratedThrift
     type FunctionType = Function1[Args,Future[Int]]
     type ReqRepFunctionType = Function1[_root_.com.twitter.scrooge.Request[Args],Future[_root_.com.twitter.scrooge.Response[Int]]]
 
-    type ServiceIfaceServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[Args, SuccessType]
     type ReqRepServicePerEndpointServiceType = _root_.com.twitter.finagle.Service[_root_.com.twitter.scrooge.Request[Args], _root_.com.twitter.scrooge.Response[SuccessType]]
-
-    def toServiceIfaceService(f: FunctionType): ServiceIfaceServiceType =
-      _root_.com.twitter.finagle.Service.mk { args: Args =>
-        f(args)
-      }
 
     def toServicePerEndpointService(f: FunctionType): ServicePerEndpointServiceType =
       _root_.com.twitter.finagle.Service.mk { args: Args =>